pyproject.toml
[tool.poetry]
name = "informed"
version = "0.1.0"
description = ""
authors = ["rahulrajesh23 <itsrahulrajesh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.2"
httpx = "^0.27.0"
openai = "^1.23.2"
pydantic = "^2.7.0"
redis = "^5.0.4"
requests = "^2.31.0"
sqlalchemy = { extras = ["asyncio", "postgresql-psycopg"], version = "^2.0.27" }
tiktoken = "^0.6.0"
uvicorn = "^0.29.0"
loguru = "^0.7.2"
opentelemetry-instrumentation-fastapi = "^0.46b0"
opentelemetry-exporter-otlp = "^1.25.0"
opentelemetry-instrumentation-logging = "^0.46b0"
pydantic-settings = "^2.5.2"
testcontainers = "^4.8.1"
sqlmodel = "^0.0.22"
alembic = "^1.13.3"
pgvector = "^0.3.4"
schedule = "^1.2.2"
# Kept but skipped in deptry checks
tenacity = "^9.0.0"
selfcheckgpt = "^0.1.7"
numpy = "1.26.4"
sentencepiece = "^0.2.0"

[tool.poetry.group.dev.dependencies]
deptry = "^0.20.0"
mypy = "^1.11.2"
pre-commit = "^3.8.0"
pyright = "^1.1.382.post1"
ruff = "^0.6.8"
black = "^24.8.0"
types-requests = "^2.32.0.20240914"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.deptry.per_rule_ignores]
DEP002 = [
    "tenacity",
    "selfcheckgpt",
    "numpy",
    "sentencepiece",
]

[tool.ruff]
exclude = ["node_modules", ".venv"]
target-version = "py311"
line-length = 120
fix = true
lint.select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E",
    "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    # "TRY",
]
lint.ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.isort]
known-third-party = ["alembic"]

[tool.mypy]
files = ["informed", "backend"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = false
show_error_codes = true
